+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Since release 0.1.0:
 - the config parameters individualEntityWhiteList and individualEntityBlackList have been removed. Use the normal white and black lists.
 - the config serialPostfix has been changed to postfix.
 
-## [0.1.3] - 2025-06-10
+## [0.1.3] - 2025-06-11
 
 ### Added
 
@@ -24,7 +24,12 @@ Since release 0.1.0:
 
 ### Changed
 
+- [package]: Update package.
+- [package]: Update dependencies.
+
 ### Fixed
+
+- [select]: Fixed ghost devices in the Device Home page.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="bmc-button.svg" alt="Buy me a coffee" width="80">